Run-clang format on the whole project
Incorporated the clang-format qt-creator (lets assume it has a valid Qt
indentation style -
https://raw.githubusercontent.com/qt-creator/qt-creator/master/.clang-format
) and then apply it to the whole code.

As a bug in clang-format - I had to move the translations into their own
dir - otherwise clang-format thinkgs this is a typescript file and will
start complaining.

Rationale: eventually - this can be a github action - and new code
cannot be pushed unless it passes clang-format.