Move some runtime code - to compile time
The compilation for OSX and Linux are different, and thus on the binaries
there is no need to have checks for "linux" on OSX.

Previous code did some checks on runtime, instead I moved this to
compile time using good old `#ifdef`. While this makes the code "ugly",
it works. I think it should be possible to make this `constexpr` and
keep a very similar look-and-feel (codewise).

As a bonus: in settings, the code tries to look for programs in $PATH,
which is "the way".